compiles now, but requires fixes for the new camera system

+++ src/core/core.cpp
@@ -280,7 +280,7 @@ void core::loadCoreConfig(const wchar_t* path) {
       coreData.at("resolution").get_to(resolution);
       config::renderWidth = resolution[0];
       config::renderHeight = resolution[1];
-      config::aspectRatio = config::renderWidth / config::renderHeight;
+      config::aspectRatio = (float)config::renderWidth / config::renderHeight;
       --requirements;
     }
